requirements.txt
opencv-python
pillow
numpy>=1.21.0,<1.25.0
pySide6
requests
psutil
gdal
#pip install "numpy>=2.0,<2.3.0"
#pip install fastapi uvicorn python-multipart aiofiles
scikit-image
scipy>=1.7.0,<2.0.0
#pip3 install torch torchvision --index-url https://download.pytorch.org/whl/cu126
tqdm
einops
fvcore
pytorchvideo
albumentations
# gdal geopandas rasterio fiona shapely
h5py
# 可视化相关依赖
matplotlib>=3.5.0
seaborn>=0.11.0